+++ connector/src/main/scala/com/microsoft/kusto/spark/utils/KustoDataSourceUtils.scala
@@ -32,6 +32,7 @@ object KustoDataSourceUtils {
   val DefaultTimeoutLongRunning: FiniteDuration = 90 minutes
   val DefaultPeriodicSamplePeriod: FiniteDuration = 2 seconds
   val NewLine = sys.props("line.separator")
+  val MaxWaitTime = 1 minute
 
   def setLoggingLevel(level: String): Unit = {
     setLoggingLevel(Level.toLevel(level))
@@ -245,7 +246,9 @@ object KustoDataSourceUtils {
   def runSequentially[A](func: () => A, delay: Int, runEvery: Int, numberOfTimesToRun: Int, doWhile: A => Boolean, finalWork: A => Unit): CountDownLatch = {
     val latch = new CountDownLatch(if (numberOfTimesToRun > 0) numberOfTimesToRun else 1)
     val t = new Timer()
-    val task = new TimerTask() {
+    var currentWaitTime = runEvery
+
+    class ExponentialBackoffTask extends  TimerTask {
       def run(): Unit = {
         val res = func.apply()
         if (numberOfTimesToRun > 0) {
@@ -257,7 +260,6 @@ object KustoDataSourceUtils {
         }
 
         if (!doWhile.apply(res)) {
-          t.cancel()
           try {
             finalWork.apply(res)
           }
@@ -267,10 +269,15 @@ object KustoDataSourceUtils {
               throw exception
           }
           while (latch.getCount > 0) latch.countDown()
+        } else {
+          currentWaitTime = if(currentWaitTime > MaxWaitTime.toMillis) currentWaitTime else currentWaitTime + currentWaitTime
+          t.schedule(new ExponentialBackoffTask() , currentWaitTime)
         }
       }
     }
-    t.schedule(task, delay, runEvery)
+
+    val task: TimerTask = new ExponentialBackoffTask()
+    t.schedule(task, delay)
     latch
   }
